ensure AuthorizeHttpClientDecorator deletes cache on ForbiddenError

+++ for_dev/lib/main/decorators/authorize_http_client_decorator.dart
@@ -29,11 +29,13 @@ class AuthorizeHttpClientDecorator implements HttpClient {
         body: body,
         headers: authorizedHeaders,
       );
-    } on HttpError {
-      rethrow;
     } catch (error) {
-      await deleteSecureCacheStorage?.deleteSecure('token');
-      throw HttpError.forbidden;
+      if (error is HttpError && error != HttpError.forbidden) {
+        rethrow;
+      } else {
+        await deleteSecureCacheStorage?.deleteSecure('token');
+        throw HttpError.forbidden;
+      }
     }
   }
 }
